% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_package_content.R
\name{package_b1}
\alias{package_b1}
\title{Thou still unravish’d bride of quietness,
Thou foster-child of silence and slow time,
Sylvan historian, who canst thus express
A flowery tale more sweetly than our rhyme:
What leaf-fring’d legend haunts about thy shape
Of deities or mortals, or of both,
In Tempe or the dales of Arcady?
What men or gods are these? What maidens loth?
What mad pursuit? What struggle to escape?
What pipes and timbrels? What wild ecstasy?}
\usage{
package_b1(a, b)
}
\arguments{
\item{a}{The best input ever}

\item{b}{The awesome second input}
}
\description{
Heard melodies are sweet, but those unheard
Are sweeter; therefore, ye soft pipes, play on;
Not to the sensual ear, but, more endear’d,
Pipe to the spirit ditties of no tone:
Fair youth, beneath the trees, thou canst not leave
Thy song, nor ever can those trees be bare;
Bold Lover, never, never canst thou kiss,
Though winning near the goal yet, do not grieve;
She cannot fade, though thou hast not thy bliss,
For ever wilt thou love, and she be fair!
Keats_urn
}
\details{
Keats’s own drawing of the Grecian Urn.

Ah, happy, happy boughs! that cannot shed
Your leaves, nor ever bid the Spring adieu;
And, happy melodist, unwearied,
For ever piping songs for ever new;
More happy love! more happy, happy love!
For ever warm and still to be enjoy’d,
For ever panting, and for ever young;
All breathing human passion far above,
That leaves a heart high-sorrowful and cloy’d,
A burning forehead, and a parching tongue.
}
